shortform-web/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
celery[redis]==5.4.0
redis==5.2.1
google-genai==1.0.0
python-dotenv==1.0.1
Pillow==11.1.0
srt==3.5.3
pydantic==2.10.4
